fix(storage): make decompression-ratio guard overflow-safe (Round 2 red-team)
Address RFL council Round 2 red-team Critical (decorrelated Gemini): the
decompression-ratio ceiling checked `uncompressedBytes > compressedFloor * ratio`
as an unchecked long*long multiply. A large declared compressed size (attacker-
controlled Parquet/checkpoint footer metadata) overflows the 64-bit product,
wrapping it to a negative or small-positive value and flipping the comparison —
either wrongly rejecting a legitimate chunk or (for a crafted size) wrongly
accepting one, defeating the eager-decode bomb guard (design §5.4 C-DECODE).

The absolute-size ceiling (ii, SaturatingAdd) still bounds total decompressed
bytes, but the ratio check itself must be correct: widen the product to Int128 so
two 64-bit operands can never overflow (exact, no truncation — unlike a divide).
This mirrors the overflow-safe intent already documented on the row-count check.

Fix both sites that share this guard: ParquetFileReader.EnsureDecodeCeiling and
DeltaCheckpointReader.ColumnFootprintBytes (the latter reaches the same defect via
the aliased ParquetFileReader.MaxDecompressionRatio constant this PR centralizes).

Regression tests: RatioGuard_LargeCompressedSize_DoesNotOverflow (a large-compressed
non-bomb chunk the pre-fix multiply wrongly rejected — proven by running the test
against the unfixed code) and RatioGuard_RealBomb_StillRejected (the widening does
not relax genuine bomb detection).

+++ src/DeltaSharp.Storage/Parquet/ParquetFileReader.cs
@@ -166,9 +166,12 @@ internal sealed class ParquetFileReader
             }
 
             // (i) Decompression-ratio ceiling: a chunk claiming far more decompressed than compressed
-            // bytes is a decompression bomb. The floor of 1 avoids a divide-by/zero-compressed edge.
+            // bytes is a decompression bomb. The floor of 1 avoids a divide-by/zero-compressed edge. The
+            // product is widened to Int128 so a large declared compressed size cannot overflow the 64-bit
+            // multiply into a spurious verdict (wrapping to a negative or small-positive threshold that would
+            // either wrongly reject a legitimate chunk or wrongly accept a bomb).
             long compressedFloor = Math.Max(chunk.CompressedBytes, 1);
-            if (chunk.UncompressedBytes > compressedFloor * limits.MaxDecompressionRatio)
+            if (chunk.UncompressedBytes > (Int128)compressedFloor * limits.MaxDecompressionRatio)
             {
                 throw DeltaStorageException.CorruptData(
                     $"Row group {group} declares {chunk.UncompressedBytes} decompressed bytes for "
